+++ al2-ruby.sh
@@ -1,5 +1,7 @@
 #! /bin/bash
 
-curl -fsSL https://github.com/rbenv/rbenv-installer/raw/master/bin/rbenv-installer | bash
-~/.rbenv/bin/rbenv install 2.5.3
-~/.rbenv/bin/rbenv global 2.5.3
+if ! [ -e ~/.rbenv/bin/rbenv ]; then
+    curl -fsSL https://github.com/rbenv/rbenv-installer/raw/master/bin/rbenv-installer | bash
+    ~/.rbenv/bin/rbenv install 2.5.3
+    ~/.rbenv/bin/rbenv global 2.5.3
+fi
