+++ ubuntu-amazon.sh
@@ -14,50 +14,60 @@ backup-and-link ~/setup/amazon.gitignore ~/.gitignore
 backup-and-link ~/setup/amazon.gitconfig ~/.gitconfig
 
 # Amazon
-wget --no-check-certificate -qO - https://cascadia.corp.amazon.com/amazon/clienteng.gpg | sudo apt-key add -
-echo deb http://cascadia.corp.amazon.com/amazon $(lsb_release -cs)-amazon main | sudo tee /etc/apt/sources.list.d/amazon.list
-echo deb http://cascadia.corp.amazon.com/amazon $(lsb_release -cs)-amazon-bh main | sudo tee -a /etc/apt/sources.list.d/amazon.list
-echo deb http://cascadia.corp.amazon.com/amazon $(lsb_release -cs)-thirdparty-partner partner | sudo tee -a /etc/apt/sources.list.d/amazon.list
-sudo apt update
-sudo apt install -y amazon-ca-certificates
+echo "--- Install Amazon repositories ---"
+if ! [ -e /etc/apt/sources.list.d/amazon.list ]; then
+    wget --no-check-certificate -qO - https://cascadia.corp.amazon.com/amazon/clienteng.gpg | sudo apt-key add -
+    echo deb http://cascadia.corp.amazon.com/amazon $(lsb_release -cs)-amazon main | sudo tee /etc/apt/sources.list.d/amazon.list
+    echo deb http://cascadia.corp.amazon.com/amazon $(lsb_release -cs)-amazon-bh main | sudo tee -a /etc/apt/sources.list.d/amazon.list
+    echo deb http://cascadia.corp.amazon.com/amazon $(lsb_release -cs)-thirdparty-partner partner | sudo tee -a /etc/apt/sources.list.d/amazon.list
+    sudo apt update
+    sudo apt install -y amazon-ca-certificates
+fi
 
 # Midway
+echo "--- Install Midway ---"
 if ! which mwinit >/dev/null 2>&1; then
     sudo apt install -y mwinit
-    mkdir -p ~/bin
-    curl https://s3.amazonaws.com/com.amazon.aws.midway.software/linux/mcurl.sh > ~/bin/mcurl
-    chmod +x ~/bin/mcurl
-    if [ -f ~/.ssh/primary.pem.pub ]; then
-        mwinit -o -k ~/.ssh/primary.pem.pub
-    else
-        test -f ~/.ssh/id_rsa.pub || ssh-keygen
-        mwinit -o
-    fi
+fi
+mkdir -p ~/bin
+curl https://s3.amazonaws.com/com.amazon.aws.midway.software/linux/mcurl.sh > ~/bin/mcurl
+chmod +x ~/bin/mcurl
+if [ -f ~/.ssh/primary.pem.pub ]; then
+    mwinit -k ~/.ssh/primary.pem.pub -l | grep -q cookie || mwinit -k ~/.ssh/primary.pem.pub -o
+else
+    test -f ~/.ssh/id_rsa.pub || ssh-keygen
+    mwinit -l | grep -q cookie || mwinit -o
 fi
 
 # Kerberos
+echo "--- Install Kerberos ---"
 if ! which kinit >/dev/null 2>&1; then
     sudo apt install -y krb5-user
     ssh-keygen -f "/home/$USER/.ssh/known_hosts" -R duybui.aka.amazon.com
     scp duybui.aka.amazon.com:/etc/krb5.conf ~
     sudo mv ~/krb5.conf /etc/krb5.conf
     sudo chown root:root /etc/krb5.conf
-    kinit -f
 fi
+klist -s || kinit -f
 
 # Toolbox
+echo "--- Install Builder Toolbox ---"
 if [ ! -e ~/.toolbox/bin/toolbox ]; then
     curl --negotiate -fLSsu: 'https://drive.corp.amazon.com/view/BuilderToolbox/toolbox-install.sh' -o /tmp/toolbox-install.sh
     bash /tmp/toolbox-install.sh
     ~/.toolbox/bin/toolbox install --channel bh toolbox
+else
+    ~/.toolbox/bin/toolbox update
 fi
 
 # CR tool
+echo "--- Install CR util ---"
 if [ ! -e ~/.toolbox/bin/cr ]; then
     ~/.toolbox/bin/toolbox install cr
 fi
 
 # Brazil 2.0
+echo "--- Install Brazil 2.0 ---"
 if [ ! -e ~/.toolbox/bin/brazil ]; then
     sudo apt install -y libreadline6
     ~/.toolbox/bin/toolbox install --channel bh brazilcli
@@ -94,11 +104,11 @@ if [ ! -e ~/.toolbox/bin/brazil ]; then
 fi
 
 # Ninja Dev Sync
-if [ ! -e ~/.toolbox/bin/ninja-dev-sync ]; then
-    curl --negotiate -fu: 'https://devcentral.amazon.com/ac/brazil/package-master/package/view/NinjaDevSync%3B2.x.2.0%3BRHEL5_64%3BDEV.STD.PTHREAD%3Bbin/ninja-dev-sync.linux64' -o ~/.toolbox/bin/ninja-dev-sync
-    chmod 755 ~/.toolbox/bin/ninja-dev-sync
-    ln -s ~/.toolbox/bin/ninja-dev-sync ~/.toolbox/bin/nds
-fi
+echo "--- Install Ninja Dev Sync ---"
+curl --negotiate -fu: 'https://devcentral.amazon.com/ac/brazil/package-master/package/view/NinjaDevSync%3B2.x.8.0%3BRHEL5_64%3BDEV.STD.PTHREAD%3Bbin/ninja-dev-sync.linux64' -o ~/.toolbox/bin/ninja-dev-sync
+chmod 755 ~/.toolbox/bin/ninja-dev-sync
+rm -f ~/.toolbox/bin/nds
+ln -s ~/.toolbox/bin/ninja-dev-sync ~/.toolbox/bin/nds
 if ! grep -q /etc/sysctl.conf -e "fs.inotify.max_user_watches"; then
     sudo apt install -y inotify-tools
     echo "fs.inotify.max_user_watches=524288" | sudo tee -a /etc/sysctl.conf
@@ -106,5 +116,6 @@ if ! grep -q /etc/sysctl.conf -e "fs.inotify.max_user_watches"; then
 fi
 
 # AWS CLI Plugins
+echo "--- Install AWS CLI plugins ---"
 ~/.pyenv/shims/pip install --user --no-warn-script-location git+ssh://git.amazon.com/pkg/BenderLibIsengard
 ~/.pyenv/shims/pip install --user --no-warn-script-location git+ssh://git.amazon.com/pkg/GoshawkBotocore@mainline-1.1
